Moved files to defs folder
